functions/related/models/default/requirements.txt
sklearn
mecab-python3
unidic-lite
